+++ Source/Core/Core/Src/DSP/DSPEmitter.cpp
@@ -10,9 +10,6 @@
 #include "DSPHost.h"
 #include "DSPInterpreter.h"
 #include "DSPAnalyzer.h"
-#include "Jit/DSPJitUtil.h"
-#include "x64Emitter.h"
-#include "x64ABI.h"
 
 #define MAX_BLOCK_SIZE 250
 #define DSP_IDLE_SKIP_CYCLES 0x1000
@@ -28,7 +25,7 @@ DSPEmitter::DSPEmitter() : gpr(*this), storeIndex(-1), storeIndex2(-1)
 	blocks = new DSPCompiledCode[MAX_BLOCKS];
 	blockLinks = new Block[MAX_BLOCKS];
 	blockSize = new u16[MAX_BLOCKS];
-	
+
 	compileSR = 0;
 	compileSR |= SR_INT_ENABLE;
 	compileSR |= SR_EXT_INT_ENABLE;
@@ -45,7 +42,7 @@ DSPEmitter::DSPEmitter() : gpr(*this), storeIndex(-1), storeIndex2(-1)
 	}
 }
 
-DSPEmitter::~DSPEmitter() 
+DSPEmitter::~DSPEmitter()
 {
 	delete[] blocks;
 	delete[] blockLinks;
@@ -65,7 +62,7 @@ void DSPEmitter::ClearIRAM()
 	g_dsp.reset_dspjit_codespace = true;
 }
 
-void DSPEmitter::ClearIRAMandDSPJITCodespaceReset() 
+void DSPEmitter::ClearIRAMandDSPJITCodespaceReset()
 {
 	ClearCodeSpace();
 	CompileDispatcher();
@@ -104,9 +101,9 @@ void DSPEmitter::checkExceptions(u32 retval)
 
 bool DSPEmitter::FlagsNeeded()
 {
-	if (!(DSPAnalyzer::code_flags[compilePC] & DSPAnalyzer::CODE_START_OF_INST) || 
-		(DSPAnalyzer::code_flags[compilePC] & DSPAnalyzer::CODE_UPDATE_SR))	
-		return true; 
+	if (!(DSPAnalyzer::code_flags[compilePC] & DSPAnalyzer::CODE_START_OF_INST) ||
+		(DSPAnalyzer::code_flags[compilePC] & DSPAnalyzer::CODE_UPDATE_SR))
+		return true;
 	else
 		return false;
 }
@@ -171,7 +168,7 @@ void DSPEmitter::EmitInstruction(UDSPInstruction inst)
 			}
 		}
 	}
-	
+
 	// Main instruction
 	if (!opTable[inst]->jitFunc)
 	{
@@ -342,7 +339,7 @@ void DSPEmitter::Compile(u16 start_addr)
 			if (!unresolvedJumps[i].empty())
 			{
 				// Check if there were any blocks waiting for this block to be linkable
-				unsigned int size = unresolvedJumps[i].size();
+				size_t size = unresolvedJumps[i].size();
 				unresolvedJumps[i].remove(start_addr);
 				if (unresolvedJumps[i].size() < size)
 				{
@@ -355,7 +352,7 @@ void DSPEmitter::Compile(u16 start_addr)
 		}
 	}
 
-	if (blockSize[start_addr] == 0) 
+	if (blockSize[start_addr] == 0)
 	{
 		// just a safeguard, should never happen anymore.
 		// if it does we might get stuck over in RunForCycles.
